// back-end/src/interface/group.interface.ts
/*Created Group Interface For saving and Updating Groups in group Table*/
export interface CreateGroupInput {
    name: string
    number_of_weeks: number
    roll_states: string
    incidents: number
    ltmt: string
    run_at: Date
    student_count: number

}

export interface UpdateGroupInput {
    id: number
    name: string
    number_of_weeks: number
    roll_states: string
    incidents: number
    ltmt: string
    run_at: Date
    student_count: number
}